implemented the search by issueKey and implemented an array for selected Issues
